move AppMigrationCommandBase to helpers.py
